+++ app/classes/models/servers.py
@@ -45,6 +45,7 @@ class Servers(BaseModel):
     ignored_exits = CharField(default="0")
     app_id = IntegerField(null=True)
     count_players = BooleanField(default=True)
+    update_watcher = BooleanField(default=False)
 
     class Meta:
         table_name = "servers"