+++ app/classes/web/routes/api/api_handlers.py
@@ -21,7 +21,10 @@ from app.classes.web.routes.api.servers.index import ApiServersIndexHandler
 from app.classes.web.routes.api.servers.server.action import (
     ApiServersServerActionHandler,
 )
-from app.classes.web.routes.api.servers.server.index import ApiServersServerIndexHandler
+from app.classes.web.routes.api.servers.server.index import (
+    ApiServersServerIndexHandler,
+    ApiServersServerUpdateConfig,
+)
 from app.classes.web.routes.api.servers.server.logs import ApiServersServerLogsHandler
 from app.classes.web.routes.api.servers.server.public import (
     ApiServersServerPublicHandler,
@@ -173,7 +176,7 @@ def api_handlers(handler_args):
             handler_args,
         ),
         (
-            r"/api/v2/crafty/JarCache/?",
+            r"/api/v2/crafty/JarCache(?:/([0-9-]+))?",
             ApiCraftyJarCacheIndexHandler,
             handler_args,
         ),
@@ -327,6 +330,11 @@ def api_handlers(handler_args):
         (
             r"/api/v2/servers/status/?",
             ApiServersServerStatusHandler,
+            handler_args,
+        ),
+        (
+            r"/api/v2/servers/([a-z0-9-]+)/update/config/?",
+            ApiServersServerUpdateConfig,
             handler_args,
         ),
         (